FEAT: Creating the video uploader [Refs #16]
