+++ core/uploader/models/__init__.py
@@ -1,0 +1,3 @@
+from .document import Document
+from .image import Image
+from .video import Video
