+++ core/uploader/serializers/__init__.py
@@ -1,0 +1,3 @@
+from .image import ImageUploadSerializer
+from .document import DocumentUploadSerializer
+from .video import VideoUploadSerializer
