+++ core/uploader/views/__init__.py
@@ -1,0 +1,3 @@
+from .image import ImageViewSet
+from .document import DocumentViewSet
+from .video import VideoViewSet
